requirements.txt
torch==2.3.1
torchvision==0.18.1
accelerate==0.33.0
transformers==4.42.4
safetensors==0.4.3
diffusers==0.30.0
runpod==1.6.2
Pillow==10.4.0
